pyproject.toml
[project]
name = "receipt-ocr-annotator"
version = "0.1.0"
description = "Expert feedback tool for OCR receipts using FastHTML, MonsterUI, and Docling"
authors = [{ name = "Your Name" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-fasthtml @ git+https://github.com/answerdotai/fasthtml@29d0cd9ed97e3529148c49240d641d460c6d971c",
    "monsterui==1.0.29",
    "docling==2.54.0",
    "uvicorn[standard]==0.37.0"
]

[tool.uv]
package = false
index-url = "https://pypi.org/simple"

[tool.uv.workspace]
members = []


